fix rest fadeloops

+++ src/about.c
@@ -183,7 +183,7 @@ void About_Unload(struct Game *game) {
 	al_draw_scaled_bitmap(game->about.letter,0,0,al_get_bitmap_width(game->about.letter),al_get_bitmap_height(game->about.letter),0,0,al_get_display_width(game->display), al_get_display_height(game->display),0);
 	al_draw_bitmap_region(game->about.text_bitmap, 0, game->about.x*al_get_bitmap_height(game->about.text_bitmap), al_get_bitmap_width(game->about.text_bitmap), al_get_display_height(game->display)*0.8, al_get_display_width(game->display)*0.5, al_get_display_height(game->display)*0.1, 0);
 	al_set_target_bitmap(al_get_backbuffer(game->display));
-	int fadeloop;
+	float fadeloop;
 	if (game->about.fadeloop!=0)
 		for(fadeloop=255; fadeloop>=0; fadeloop-=tps(game, 600)){
 			al_wait_for_event(game->event_queue, &ev);
