+++ src/intro.c
@@ -167,7 +167,7 @@ void Intro_Draw(struct Game *game) {
 void Intro_Load(struct Game *game) {
 	al_play_sample_instance(game->intro.music);
 	ALLEGRO_EVENT ev;
-	int fadeloop;
+	float fadeloop;
 	for(fadeloop=0; fadeloop<256; fadeloop+=tps(game, 600)){
 		al_wait_for_event(game->event_queue, &ev);
 		al_draw_tinted_bitmap(game->intro.table,al_map_rgba_f(fadeloop/255.0,fadeloop/255.0,fadeloop/255.0,1),0,0,0);
@@ -243,7 +243,7 @@ void Intro_Unload(struct Game *game) {
 		al_destroy_audio_stream(game->intro.audiostream);
 	}
 	ALLEGRO_EVENT ev;
-	int fadeloop;
+	float fadeloop;
 	for(fadeloop=255; fadeloop>=0; fadeloop-=tps(game, 600)){
 		al_wait_for_event(game->event_queue, &ev);
 		if (game->intro.in_animation) {
