+++ src/map.c
@@ -59,7 +59,7 @@ void Map_Draw(struct Game *game) {
 void Map_Load(struct Game *game) {
 	al_play_sample_instance(game->map.music);
 	ALLEGRO_EVENT ev;
-	int fadeloop;
+	float fadeloop;
 	for(fadeloop=0; fadeloop<256; fadeloop+=tps(game, 600)){
 		al_wait_for_event(game->event_queue, &ev);
 		al_draw_tinted_bitmap(game->map.map,al_map_rgba_f(fadeloop/255.0,fadeloop/255.0,fadeloop/255.0,1),0,0,0);
@@ -140,7 +140,7 @@ void Map_Preload(struct Game *game) {
 
 void Map_Unload(struct Game *game) {
 	ALLEGRO_EVENT ev;
-	int fadeloop;
+	float fadeloop;
 	for(fadeloop=255; fadeloop>=0; fadeloop-=tps(game, 600)){
 		al_wait_for_event(game->event_queue, &ev);
 		al_draw_tinted_bitmap(game->map.map, al_map_rgba_f(fadeloop/255.0,fadeloop/255.0,fadeloop/255.0,1), 0, 0, 0);
